cambios tablero cliente
se modifica endpoints para poder guarrdar editar o inactivar clientes eb tablero de clientes

+++ Voalaft.Data/Entidades/CatClientes.cs
@@ -30,7 +30,7 @@ namespace Voalaft.Data.Entidades
             
         public bool bManejaCredito { get; set; }
 
-        public byte? nTipoPersona { get; set; }
+        public int? nTipoPersona { get; set; }
 
         public long? nIDRFC { get; set; }
